+++ app/src/main/java/de/christian2003/chaching/plugin/presentation/widget/OverviewWidget.kt
@@ -1,7 +1,6 @@
 package de.christian2003.chaching.plugin.presentation.widget
 
 import android.content.Context
-import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
@@ -14,7 +13,6 @@ import androidx.glance.Image
 import androidx.glance.ImageProvider
 import androidx.glance.LocalContext
 import androidx.glance.LocalSize
-import androidx.glance.appwidget.CircularProgressIndicator
 import androidx.glance.appwidget.GlanceAppWidget
 import androidx.glance.appwidget.SizeMode
 import androidx.glance.appwidget.cornerRadius
@@ -45,19 +43,30 @@ import java.text.NumberFormat
 import java.time.LocalDate
 
 
+/**
+ * Implements the overview widget which displays the total amount of money earned in the last 31
+ * days.
+ */
 class OverviewWidget : GlanceAppWidget() {
 
+    /**
+     * Number format for the value displayed.
+     */
     private val numberFormat: NumberFormat = DecimalFormat("#,###.00")
 
 
-    companion object {
-        private val SMALL_SQUARE = DpSize(140.dp, 70.dp)
-        private val LARGE_SQUARE = DpSize(width = 210.dp, height = 70.dp)
-    }
-
+    /**
+     * Returns the size modes that are supported by the widget implementation.
+     */
     override val sizeMode: SizeMode = SizeMode.Responsive(setOf(SMALL_SQUARE, LARGE_SQUARE))
 
 
+    /**
+     * Provides the Android Glance code for the widget.
+     *
+     * @param context   Context for the widget.
+     * @param id        ID of the widget.
+     */
     override suspend fun provideGlance(context: Context, id: GlanceId) {
         var isError = false
         var data: OverviewCalcResult? = null
@@ -69,7 +78,7 @@ class OverviewWidget : GlanceAppWidget() {
                 val transfers: List<Transfer> = repository.getAllTransfersInDateRange(now.minusDays(31), now).first()
                 data = OverviewCalcResult(transfers, types)
             }
-            catch (e: Exception) {
+            catch (_: Exception) {
                 isError = true
             }
             finally {
@@ -83,11 +92,7 @@ class OverviewWidget : GlanceAppWidget() {
                     //Data cannot be loaded:
                     ErrorDisplay()
                 }
-                else if (data == null) {
-                    //Data is loading:
-                    LoadingDisplay()
-                }
-                else if (data.totalValue == 0) {
+                else if (data == null || data.totalValue == 0) {
                     //No incomes:
                     EmptyDisplay()
                 }
@@ -100,6 +105,9 @@ class OverviewWidget : GlanceAppWidget() {
     }
 
 
+    /**
+     * Displays the normal state. This is displayed once the data is loaded successfully.
+     */
     @Composable
     private fun OverviewDisplay(data: OverviewCalcResult) {
         val size = LocalSize.current
@@ -150,6 +158,12 @@ class OverviewWidget : GlanceAppWidget() {
     }
 
 
+    /**
+     * Displays the value earned in the last 31 days.
+     *
+     * @param value     Value in cents to display.
+     * @param modifier  Glance modifier.
+     */
     @Composable
     private fun ValueDisplay(
         value: Int,
@@ -174,20 +188,10 @@ class OverviewWidget : GlanceAppWidget() {
     }
 
 
-    @Composable
-    private fun LoadingDisplay() {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = GlanceModifier
-                .fillMaxSize()
-                .background(GlanceTheme.colors.surface)
-        ) {
-            CircularProgressIndicator()
-        }
-    }
-
-
+    /**
+     * Displays the empty state. This is displayed if the user has not earned any money in the
+     * last 31 days.
+     */
     @Composable
     private fun EmptyDisplay() {
         val size = LocalSize.current
@@ -227,6 +231,9 @@ class OverviewWidget : GlanceAppWidget() {
     }
 
 
+    /**
+     * Displays the error state. This is displayed if some error occurs while loading data.
+     */
     @Composable
     private fun ErrorDisplay() {
         val size = LocalSize.current
@@ -263,6 +270,21 @@ class OverviewWidget : GlanceAppWidget() {
                 }
             }
         }
+    }
+
+
+    companion object {
+
+        /**
+         * Sizes for the small widget (2 x 1 cells).
+         */
+        private val SMALL_SQUARE = DpSize(140.dp, 70.dp)
+
+        /**
+         * Sizes for the large widget (3 x 1 cells).
+         */
+        private val LARGE_SQUARE = DpSize(width = 210.dp, height = 70.dp)
+
     }
 
 }